successfully split everythying into libaries
