+++ CMakeLists.txt
@@ -1,17 +1,22 @@
-cmake_minimum_required(VERSION 3.16)
-project(abeams LANGUAGES CXX)
+cmake_minimum_required(VERSION 3.20)
+
+set(CMAKE_TOOLCHAIN_FILE "${CMAKE_CURRENT_SOURCE_DIR}/vcpkg/scripts/buildsystems/vcpkg.cmake"
+        CACHE STRING "Vcpkg toolchain file")
+
+project(abeams)
 
 set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
-
 find_package(fmt CONFIG REQUIRED)
-find_package(nlohmann_json CONFIG REQUIRED)
 find_package(spdlog CONFIG REQUIRED)
-find_package(boost CONFIG REQUIRED COMPONENTS asio beast)
+find_package(nlohmann_json CONFIG REQUIRED)
 
-
-add_subdirectory(core)
-add_subdirectory(net)
 add_subdirectory(util)
+add_subdirectory(net)
 add_subdirectory(cli)
+add_subdirectory(core)
+
+if(WIN32)
+    target_link_libraries(abeams PRIVATE ws2_32 mswsock)
+endif()
