+++ core/CMakeLists.txt
@@ -1,10 +1,19 @@
-add_executable(engine
-    main.cpp
-    initialize.h
+
+add_subdirectory(sim)
+add_subdirectory(thread)
+
+add_executable(abeams
+        main.cpp
+        initialize.h
 )
 
-target_link_libraries(engine
-    PRIVATE
+target_include_directories(abeams
+        PRIVATE
+        ${CMAKE_SOURCE_DIR}
+)
+
+target_link_libraries(abeams
+        PRIVATE
         abeams_thread
         abeams_util
         abeams_sim
@@ -13,6 +22,3 @@ target_link_libraries(engine
         spdlog::spdlog
 )
 
-if(WIN32)
-    target_link_libraries(engine PRIVATE ws2_32 mswsock)
-endif()
