+++ core/sim/sim.h
@@ -1,0 +1,16 @@
+//
+// Created by theow on 03.09.2025.
+//
+
+#ifndef SIM_H
+#define SIM_H
+
+
+
+class sim {
+
+};
+
+
+
+#endif //SIM_H
