+++ core/thread/CMakeLists.txt
@@ -1,12 +1,11 @@
-add_library(abeams_thread SHARED
-    threadpool/threadpool.cpp
-    threadpool/threadpool.h
+
+add_library(abeams_thread STATIC
+        threadpool/threadpool.cpp
+        threadpool/threadpool.h
 )
 
 target_include_directories(abeams_thread
-    PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}
+        PUBLIC
+        ${CMAKE_SOURCE_DIR}
 )
 
-target_link_libraries(abeams_thread
-    PUBLIC fmt::fmt
-)
