+++ net/CMakeLists.txt
@@ -1,4 +1,5 @@
-add_library(abeams_net SHARED
+
+add_library(abeams_net STATIC
         websockets/wsServer.cpp
         websockets/wsServer.h
         websockets/wsHandler.cpp
@@ -7,21 +8,13 @@ add_library(abeams_net SHARED
 
 target_include_directories(abeams_net
         PUBLIC
-        ${CMAKE_CURRENT_SOURCE_DIR}
-        ${CMAKE_CURRENT_SOURCE_DIR}/../header
+        ${CMAKE_SOURCE_DIR}
 )
 
 target_link_libraries(abeams_net
-        PRIVATE
+        PUBLIC
+        abeams_util
         fmt::fmt
         nlohmann_json::nlohmann_json
 )
 
-if(WIN32)
-    target_link_libraries(abeams_net PRIVATE ws2_32 mswsock)
-endif()
-
-target_include_directories(abeams_net
-        PUBLIC
-        ${CMAKE_SOURCE_DIR}/vcpkg/installed/x64-windows/include
-)
