+++ util/CMakeLists.txt
@@ -1,20 +1,21 @@
-add_library(abeams_util SHARED
-    parseConfig.cpp
-    parseConfig.h
-    logsys/logsys.cpp
-    logsys/logsys.h
-    logsys/logToFile.cpp
-    logsys/logToFile.h
+add_library(abeams_util STATIC
+        parseConfig.cpp
+        parseConfig.h
+        logsys/logsys.cpp
+        logsys/logsys.h
+        logsys/logToFile.cpp
+        logsys/logToFile.h
 )
 
 target_include_directories(abeams_util
-    PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}
+        PUBLIC
+        ${CMAKE_SOURCE_DIR}
 )
 
 target_link_libraries(abeams_util
-    PUBLIC
+        PUBLIC
         fmt::fmt
         nlohmann_json::nlohmann_json
-    PRIVATE
+        PRIVATE
         spdlog::spdlog
 )
